Revert "reformat all"
This reverts commit 88a08c12318b3607d6cd95fa86e17e4b7191091b.

+++ src/main/java/net/bhl/matsim/uam/analysis/traffic/CSVLinkStatsWriter.java
@@ -4,7 +4,11 @@ import java.io.BufferedWriter;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.SortedSet;
+import java.util.TreeSet;
 
 public class CSVLinkStatsWriter {
 	final private Collection<LinkStatsItem> links;
@@ -34,7 +38,7 @@ public class CSVLinkStatsWriter {
 		header.add("link_id");
 		header.add("length_m");
 		header.add("freespeed_ms");
-
+		
 		timeHeaders = new TreeSet<Integer>();
 		for (LinkStatsItem link : links) {
 			for (Integer i : link.timeDependantSpeed.keySet()) {
@@ -43,7 +47,7 @@ public class CSVLinkStatsWriter {
 		}
 
 		for (Integer timeHead : timeHeaders) {
-			header.add("avgspeed_ms_at_H" + (timeHead / 3600) + "M" + ((timeHead % 3600) / 60));
+			header.add("avgspeed_ms_at_H" + (int) (timeHead/ 3600) + "M" + (int) ((timeHead % 3600) / 60));
 		}
 
 		return String.join(delimiter, header);
@@ -54,7 +58,7 @@ public class CSVLinkStatsWriter {
 		row.add(link.linkId.toString());
 		row.add(String.valueOf(link.distance));
 		row.add(String.valueOf(link.freeSpeed));
-
+		
 		for (Integer timeHead : timeHeaders) {
 			if (link.timeDependantSpeed.containsKey(timeHead))
 				row.add(String.valueOf(link.timeDependantSpeed.get(timeHead)));
